# backend/src/api/api.yaml
openapi: 3.0.0
info:
  title: Senior Design Project API
  description: API for managing beam data, machines, results, baselines, and thresholds
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: https://staging-api.example.com/v1
    description: Staging server
paths:
...
  /updates:
    get:
      tags:
        - Updates
      summary: Get all updates
      description: Returns a list of all updates
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Update'
    
    post:
      tags:
        - Updates
      summary: Create a new update
      description: Adds a new update entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Update'
      responses:
        '201':
          description: Update created successfully
        '400':
          description: Bad request
    
    put:
      tags:
        - Updates
      summary: Update an existing update
      description: Modifies an existing update entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Update'
      responses:
        '200':
          description: Update modified successfully
        '400':
          description: Bad request
        '404':
          description: Update not found
    
    delete:
      tags:
        - Updates
      summary: Delete an update
      description: Removes an update entry
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Update identifier
      responses:
        '204':
          description: Update deleted successfully
        '404':
          description: Update not found
...
  /machines:
    get:
      tags:
        - Machines
      summary: Get all machines
      description: Returns a list of all machines
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Machine'

    post:
      tags:
        - Machines
      summary: Create a machine
      description: Creates a new machine entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Machine'
      responses:
        '201':
          description: Machine created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '409':
          description: Conflict - machine already exists
  /machines/{id}:
    get:
      tags:
        - Machines
      summary: Get machine by ID
      description: Retrieves a single machine by its identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique machine identifier
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Machine'
        '404':
          description: Machine not found

    put:
      tags:
        - Machines
      summary: Update machine
      description: Updates an existing machine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique machine identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Machine'
      responses:
        '204':
          description: Machine updated successfully
        '400':
          description: Bad request - ID mismatch
        '404':
          description: Machine not found

    delete:
      tags:
        - Machines
      summary: Delete machine
      description: Removes a machine from the system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique machine identifier
      responses:
        '204':
          description: Machine deleted successfully
        '404':
          description: Machine not found
...
  /updates/{id}:
    get:
      tags:
        - Updates
      summary: Get update by ID
      description: Retrieves a single update by its identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique update identifier
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Update'
        '404':
          description: Update not found
  /geochecks:
    get:
      tags:
        - Geocheck
      summary: Get geometry checks
      description: Retrieves geometry check data, optionally filtered by machine, type, and date range
      parameters:
        - name: machine-id
          in: query
          required: false
          schema:
            type: string
          description: Machine identifier
        - name: type
          in: query
          required: false
          schema:
            type: string
          description: Type of beam (e.g., 6xff)
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Specific date for geometry check
        - name: start-date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Start date for date range
        - name: end-date
          in: query
          required: false
          schema:
            type: string
            format: date
          description: End date for date range
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeoCheck'

    post:
      tags:
        - Geocheck
      summary: Create a geometry check
      description: Creates a new geometry check entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoCheck'
      responses:
        '201':
          description: Geometry check created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoCheck'
        '400':
          description: Bad request
  /geochecks/{id}:
    get:
      tags:
        - Geocheck
      summary: Get geometry check by ID
      description: Retrieves a single geometry check by its identifier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique geometry check identifier
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoCheck'
        '404':
          description: Geometry check not found

    put:
      tags:
        - Geocheck
      summary: Update geometry check
      description: Updates an existing geometry check
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique geometry check identifier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeoCheck'
      responses:
        '200':
          description: Geometry check updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoCheck'
        '400':
          description: Bad request
        '404':
          description: Geometry check not found

    delete:
      tags:
        - Geocheck
      summary: Delete geometry check
      description: Removes a geometry check from the system
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Unique geometry check identifier
      responses:
        '204':
          description: Geometry check deleted successfully
        '404':
          description: Geometry check not found
components:
  schemas:
...
    Machine:
      type: object
      properties:
        Id:
          type: string
          description: Unique identifier for the machine
        Location:
          type: string
          description: Physical location of the machine
        Name:
          type: string
          description: Name of the machine
        Type:
          type: string
          description: Type of machine
      required:
        - Id
        - Location
        - Name
        - Type
...
    GeoCheck:
      type: object
      properties:
        Id:
          type: string
          description: Unique identifier for the geometry check
        Type:
          type: string
          description: Type of beam (e.g., 6xff)
        Date:
          type: string
          format: date
          description: Date of the geometry check
        MachineId:
          type: string
          description: Associated machine identifier
        Path:
          type: string
          description: File path to the geometry check data
        IsoCenterSize:
          type: number
          format: double
          description: Iso center size measurement
        IsoCenterMVOffset:
          type: number
          format: double
          description: Iso center MV offset measurement
        IsoCenterKVOffset:
          type: number
          format: double
          description: Iso center KV offset measurement
        RelativeOutput:
          type: number
          format: double
          description: Relative output value
        RelativeUniformity:
          type: number
          format: double
          description: Relative uniformity value
        CenterShift:
          type: number
          format: double
          description: Center shift value
        CollimationRotationOffset:
          type: number
          format: double
          description: Collimation rotation offset measurement
        GantryAbsolute:
          type: number
          format: double
          description: Gantry absolute measurement
        GantryRelative:
          type: number
          format: double
          description: Gantry relative measurement
        CouchMaxPositionError:
          type: number
          format: double
          description: Maximum couch position error
        CouchLat:
          type: number
          format: double
          description: Couch lateral measurement
        CouchLng:
          type: number
          format: double
          description: Couch longitudinal measurement
        CouchVrt:
          type: number
          format: double
          description: Couch vertical measurement
        CouchRtnFine:
          type: number
          format: double
          description: Couch rotation fine measurement
        CouchRtnLarge:
          type: number
          format: double
          description: Couch rotation large measurement
        RotationInducedCouchShiftFullRange:
          type: number
          format: double
          description: Rotation induced couch shift full range
        MaxOffsetA:
          type: number
          format: double
          description: Maximum offset for MLC bank A
        MaxOffsetB:
          type: number
          format: double
          description: Maximum offset for MLC bank B
        MeanOffsetA:
          type: number
          format: double
          description: Mean offset for MLC bank A
        MeanOffsetB:
          type: number
          format: double
          description: Mean offset for MLC bank B
        MlcBacklashMaxA:
          type: number
          format: double
          description: Maximum backlash for MLC bank A
        MlcBacklashMaxB:
          type: number
          format: double
          description: Maximum backlash for MLC bank B
        MlcBacklashMeanA:
          type: number
          format: double
          description: Mean backlash for MLC bank A
        MlcBacklashMeanB:
          type: number
          format: double
          description: Mean backlash for MLC bank B
        JawX1:
          type: number
          format: double
          description: Jaw X1 measurement
        JawX2:
          type: number
          format: double
          description: Jaw X2 measurement
        JawY1:
          type: number
          format: double
          description: Jaw Y1 measurement
        JawY2:
          type: number
          format: double
          description: Jaw Y2 measurement
        JawParallelismX1:
          type: number
          format: double
          description: Jaw parallelism X1 measurement
        JawParallelismX2:
          type: number
          format: double
          description: Jaw parallelism X2 measurement
        JawParallelismY1:
          type: number
          format: double
          description: Jaw parallelism Y1 measurement
        JawParallelismY2:
          type: number
          format: double
          description: Jaw parallelism Y2 measurement
        MlcLeavesA:
          type: object
          additionalProperties:
            type: number
            format: double
          description: MLC leaves bank A measurements keyed by leaf label (Leaf11-Leaf50)
        MlcLeavesB:
          type: object
          additionalProperties:
            type: number
            format: double
          description: MLC leaves bank B measurements keyed by leaf label (Leaf11-Leaf50)
        MlcBacklashA:
          type: object
          additionalProperties:
            type: number
            format: double
          description: MLC backlash bank A measurements keyed by leaf label (Leaf11-Leaf50)
        MlcBacklashB:
          type: object
          additionalProperties:
            type: number
            format: double
          description: MLC backlash bank B measurements keyed by leaf label (Leaf11-Leaf50)
        Note:
          type: string
          description: Notes about the geometry check
      required:
        - Id
        - Type
        - Date
        - MachineId
...
    Update:
      type: object
      properties:
        Id:
          type: string
          description: Unique identifier for the update
        MachineId:
          type: string
          description: Associated machine
        Info:
          type: string
          description: Update information
        Type:
          type: string
          description: Type of update
      required:
        - Id
        - MachineId
        - Info
        - Type
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []